* merge back changes from VST SDK 3.5:
		* fix some warnings with some compilers
		* add CAnimationSplashScreen class
		* animations can have an optional notification object for receiving a message when the animation has finished

+++ vstgui/lib/cdropsource.cpp
@@ -81,7 +81,7 @@ bool CDropSource::add (const void* buffer, int32_t bufferSize, Type type)
 //-----------------------------------------------------------------------------
 int32_t CDropSource::getCount () const
 {
-	return entries.size ();
+	return (int32_t)entries.size ();
 }
 
 //-----------------------------------------------------------------------------
